WEEK9/ecommerce_project/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-filter>=23.0
psycopg2-binary>=2.9   # Optional — only needed for PostgreSQL
